Added storing of the current visualization and synthesis of that file.
Via the "Store View Model" action the current visualization context is
now stored on disc in the relative path
models/[projectName]-visualization-[yyyyMMddHHmmss].osgiviz
as the XMI variant of the underlying EMF model. The view model also
stores the current osgimodel with it in the same file. That model is
also viewable by KLighD with a new synthesis resulting in an equivalent
view as if the osgimodel was viewed with that configuration.